+++ examples/runimage-alpine-busybox.sh
@@ -22,7 +22,7 @@ echo -e 'options timeout:2 attempts:3/nnameserver 1.1.1.1\nnameserver 8.8.8.8' >
 
 RIM_ROOT=1 ./runimage apk add bash coreutils curl findutils gawk grep iproute2 kmod procps-ng sed \
 tar util-linux which gocryptfs libnotify lsof slirp4netns socat xhost gzip xz zstd lz4 jq binutils \
-patchelf nftables iptables openresolv iputils file fakeroot||true
+patchelf nftables iptables openresolv iputils file fakeroot dbus||true
 
 curl -L https://raw.githubusercontent.com/VHSgunzo/runimage-fake-sudo-pkexec/refs/heads/main/usr/bin/sudo \
     -o rootfs/usr/bin/sudo && chmod +x rootfs/usr/bin/sudo