+++ examples/runimage-debian.sh
@@ -18,7 +18,7 @@ echo -e 'nameserver 1.1.1.1\nnameserver 8.8.8.8' > rootfs/etc/resolv.conf
 
 RIM_ROOT=1 ./runimage sh -c 'apt update && apt install -y bash coreutils curl findutils gawk grep iproute2 \
 kmod procps sed tar util-linux gnu-which gocryptfs libnotify-bin lsof slirp4netns socat x11-xserver-utils \
-gzip xz-utils zstd lz4 jq binutils patchelf nftables iptables openresolv iputils-ping fakeroot fakechroot file'||true
+gzip xz-utils zstd lz4 jq binutils patchelf nftables iptables openresolv iputils-ping fakeroot fakechroot file dbus'||true
 
 curl -L https://raw.githubusercontent.com/VHSgunzo/runimage-fake-sudo-pkexec/refs/heads/main/usr/bin/sudo \
     -o rootfs/usr/bin/sudo && chmod +x rootfs/usr/bin/sudo