+++ examples/runimage-void.sh
@@ -19,7 +19,7 @@ echo -e 'nameserver 1.1.1.1\nnameserver 8.8.8.8' > rootfs/etc/resolv.conf
 RIM_ROOT=1 ./runimage sh -c 'xbps-install -Syu xbps ; \
 xbps-install -Sy bash coreutils curl findutils gawk grep iproute2 kmod \
 procps-ng sed tar util-linux which gocryptfs libnotify lsof slirp4netns socat xhost gzip xz \
-zstd lz4 jq binutils patchelf nftables iptables-nft openresolv iputils fakeroot file'||true
+zstd lz4 jq binutils patchelf nftables iptables-nft openresolv iputils fakeroot file dbus ; ldconfig'||true
 
 curl -L https://raw.githubusercontent.com/VHSgunzo/runimage-fake-sudo-pkexec/refs/heads/main/usr/bin/sudo \
     -o rootfs/usr/bin/sudo && chmod +x rootfs/usr/bin/sudo